Fix upload flow
